// src/app/routes/courses/model/course-interface.ts
import { StringLiteral } from 'typescript';

export interface CourseModel {
    courseCode: string;
    courseId: number;
    courseName: string;
    courseStream: string;
    courseType: string;
    courseYear: number;
    status: boolean;
    iconPath: string;
    masCourseType: string;
    masCourseYear: any;
    masStreamCode: string;
    boardId: number;
    courseTypeId: number;
    masStream: any;
    boardName: string;
}